长期记忆保存功能完善

+++ companion-ai/db/models.py
@@ -13,8 +13,8 @@ CREATE TABLE IF NOT EXISTS users (
     username    TEXT,
     nickname    TEXT,
     role_id     TEXT    NOT NULL DEFAULT 'Alex',
-    created_at  REAL    NOT NULL DEFAULT (unixepoch()),
-    last_active_at REAL NOT NULL DEFAULT (unixepoch())
+    created_at  INTEGER    NOT NULL DEFAULT (unixepoch()),
+    last_active_at INTEGER NOT NULL DEFAULT (unixepoch())
 );
 
 CREATE TABLE IF NOT EXISTS conversations (
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS conversations (
     role        TEXT    NOT NULL,   -- 'user' or 'assistant'
     content     TEXT    NOT NULL,
     emotion_tag TEXT,
-    created_at  REAL    NOT NULL DEFAULT (unixepoch())
+    created_at  INTEGER    NOT NULL DEFAULT (unixepoch())
 );
 
 CREATE TABLE IF NOT EXISTS memories (
@@ -31,7 +31,7 @@ CREATE TABLE IF NOT EXISTS memories (
     user_id     INTEGER NOT NULL,
     content     TEXT    NOT NULL,
     importance  INTEGER NOT NULL DEFAULT 1,
-    created_at  REAL    NOT NULL DEFAULT (unixepoch())
+    created_at  INTEGER NOT NULL DEFAULT (unixepoch())
 );
 
 CREATE TABLE IF NOT EXISTS long_term_memories (
@@ -42,10 +42,10 @@ CREATE TABLE IF NOT EXISTS long_term_memories (
     keywords_json   TEXT    NOT NULL DEFAULT '[]',
     confidence      REAL    NOT NULL DEFAULT 0.8,
     status          TEXT    NOT NULL DEFAULT 'active',
-    happened_at     REAL,
-    created_at      REAL    NOT NULL DEFAULT (unixepoch()),
-    updated_at      REAL    NOT NULL DEFAULT (unixepoch()),
-    last_seen_at    REAL    NOT NULL DEFAULT (unixepoch())
+    happened_at     INTEGER,
+    created_at      INTEGER    NOT NULL DEFAULT (unixepoch()),
+    updated_at      INTEGER    NOT NULL DEFAULT (unixepoch()),
+    last_seen_at    INTEGER    NOT NULL DEFAULT (unixepoch())
 );
 
 CREATE INDEX IF NOT EXISTS idx_ltm_user_id
@@ -61,7 +61,7 @@ CREATE TABLE IF NOT EXISTS scheduler_tasks (
     id          INTEGER PRIMARY KEY AUTOINCREMENT,
     user_id     INTEGER NOT NULL,
     task_type   TEXT    NOT NULL,
-    trigger_at  REAL    NOT NULL,
+    trigger_at  INTEGER    NOT NULL,
     status      TEXT    NOT NULL DEFAULT 'pending'
 );
 """
